Delete webgpu support, move everything to ops_webgpu
